worked on admin page frontend

+++ backend/src/main/java/com/basics/backend/dto/ProductDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ProductDto {
 
     private Long productId;
-    private Double quantity;
+    private Long quantity;
 }
